// apps/dashboard-api/src/modules/hikvision/dto/auth-mode.enum.ts
export enum AuthMode {
    FACE = 'FACE',
    CARD = 'CARD',
    PASSWORD = 'PASSWORD',
    CARD_OR_FACE_OR_PASSWORD = 'CARD_OR_FACE_OR_PASSWORD',
    FACE_AND_CARD = 'FACE_AND_CARD',
    FACE_AND_PASSWORD = 'FACE_AND_PASSWORD',
    CARD_AND_PASSWORD = 'CARD_AND_PASSWORD',
    FINGERPRINT = 'FINGERPRINT',
    FINGERPRINT_OR_CARD = 'FINGERPRINT_OR_CARD',
    FINGERPRINT_AND_CARD = 'FINGERPRINT_AND_CARD',
    CARD_OR_FACE = 'CARD_OR_FACE',
    CARD_OR_FINGERPRINT_OR_PASSWORD = 'CARD_OR_FINGERPRINT_OR_PASSWORD',
}